+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/events/KubernetesRestartEventPublisher.java
@@ -98,23 +98,9 @@ public class KubernetesRestartEventPublisher {
         ObjectReference resourceReference = createResourceReference(reconciliation);
 
         try {
-            EventBuilder builder = new EventBuilder();
-
-            builder.withNewMetadata()
-                    .withGenerateName("strimzi-event")
-                    .endMetadata()
-                    .withAction(action)
-                    .withReportingController(CONTROLLER)
-                    .withReportingInstance(operatorName)
-                    .withRegarding(resourceReference)
-                    .withReason(reason)
-                    .withType(type)
-                    .withEventTime(k8sEventTime)
-                    .withNote(maybeTruncated(note));
-
             LOG.debug("Publishing K8s event, time={}, type={}, reason={}, note={}, resource={}",
                     k8sEventTime, type, reason, note, resourceReference);
-            client.events().v1().events().inNamespace(reconciliation.namespace()).resource(builder.build()).create();
+            createAndPublishEvent(k8sEventTime, resourceReference, null, action, reason, type, maybeTruncated(note));
         } catch (Exception e) {
             LOG.error("Exception on K8s event publication", e);
         }
@@ -131,12 +117,16 @@ public class KubernetesRestartEventPublisher {
      * @param note              - the note to attach to the event
      */
     protected void publishEvent(MicroTime eventTime, ObjectReference resourceReference, ObjectReference podReference, String reason, String type, String note) {
+        createAndPublishEvent(eventTime, resourceReference, podReference, ACTION, reason, type, note);
+    }
+
+    private void createAndPublishEvent(MicroTime eventTime, ObjectReference resourceReference, ObjectReference podReference, String action, String reason, String type, String note) {
         EventBuilder builder = new EventBuilder();
 
         builder.withNewMetadata()
                 .withGenerateName("strimzi-event")
                 .endMetadata()
-                .withAction(ACTION)
+                .withAction(action)
                 .withReportingController(CONTROLLER)
                 .withReportingInstance(operatorName)
                 .withRegarding(resourceReference)
@@ -146,7 +136,7 @@ public class KubernetesRestartEventPublisher {
                 .withEventTime(eventTime)
                 .withNote(note);
 
-        client.events().v1().events().inNamespace(podReference.getNamespace()).resource(builder.build()).create();
+        client.events().v1().events().inNamespace(resourceReference.getNamespace()).resource(builder.build()).create();
     }
 
     ObjectReference createPodReference(Pod pod) {